+++ controller/api/destination/external-workload/endpoints_controller.go
@@ -7,7 +7,7 @@ import (
 	"sync"
 	"time"
 
-	ewv1alpha1 "github.com/linkerd/linkerd2/controller/gen/apis/externalworkload/v1alpha1"
+	ewv1beta1 "github.com/linkerd/linkerd2/controller/gen/apis/externalworkload/v1beta1"
 	"github.com/linkerd/linkerd2/controller/k8s"
 	logging "github.com/sirupsen/logrus"
 	corev1 "k8s.io/api/core/v1"
@@ -62,6 +62,7 @@ type EndpointsController struct {
 
 	lec leaderelection.LeaderElectionConfig
 	informerHandlers
+	dropsMetric workqueue.CounterMetric
 }
 
 // informerHandlers holds handles to callbacks that have been registered with
@@ -93,11 +94,16 @@ type informerHandlers struct {
 // NewEndpointsController creates a new controller. The controller must be
 // started with its `Start()` method.
 func NewEndpointsController(k8sAPI *k8s.API, hostname, controllerNs string, stopCh chan struct{}, exportQueueMetrics bool) (*EndpointsController, error) {
+	queueName := "endpoints_controller_workqueue"
 	workQueueConfig := workqueue.RateLimitingQueueConfig{
-		Name: "endpoints_controller_workqueue",
+		Name: queueName,
 	}
+
+	var dropsMetric workqueue.CounterMetric = &noopCounterMetric{}
 	if exportQueueMetrics {
-		workQueueConfig.MetricsProvider = newWorkQueueMetricsProvider()
+		provider := newWorkQueueMetricsProvider()
+		workQueueConfig.MetricsProvider = provider
+		dropsMetric = provider.NewDropsMetric(queueName)
 	}
 
 	ec := &EndpointsController{
@@ -108,6 +114,7 @@ func NewEndpointsController(k8sAPI *k8s.API, hostname, controllerNs string, stop
 		log: logging.WithFields(logging.Fields{
 			"component": "external-endpoints-controller",
 		}),
+		dropsMetric: dropsMetric,
 	}
 
 	// Store configuration for leader elector client. The leader elector will
@@ -333,6 +340,7 @@ func (ec *EndpointsController) handleError(err error, key string) {
 	}
 
 	ec.queue.Forget(key)
+	ec.dropsMetric.Inc()
 	ec.log.Errorf("dropped Service %s out of update queue: %v", key, err)
 }
 
@@ -389,7 +397,7 @@ func (ec *EndpointsController) syncService(update string) error {
 
 	epSlices = dropEndpointSlicesPendingDeletion(epSlices)
 	if ec.reconciler.endpointTracker.StaleSlices(svc, epSlices) {
-		ec.log.Warnf("detected EndpointSlice informer cache is out of date when processing %s", svc)
+		ec.log.Warnf("detected EndpointSlice informer cache is out of date when processing %s", update)
 		return errors.New("EndpointSlice informer cache is out of date")
 	}
 	err = ec.reconciler.reconcile(svc, ews, epSlices)
@@ -494,7 +502,7 @@ func (ec *EndpointsController) queueServiceForEndpointSlice(endpointSlice *disco
 }
 
 func (ec *EndpointsController) onAddExternalWorkload(obj interface{}) {
-	ew, ok := obj.(*ewv1alpha1.ExternalWorkload)
+	ew, ok := obj.(*ewv1beta1.ExternalWorkload)
 	if !ok {
 		ec.log.Errorf("couldn't get ExternalWorkload from object %#v", obj)
 		return